+++ src/main/java/me/aekrylov/technaxis_test/BookRepository.java
@@ -1,6 +1,5 @@
 package me.aekrylov.technaxis_test;
 
-import me.aekrylov.technaxis_test.Book;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Modifying;
@@ -11,11 +10,11 @@ import org.springframework.data.repository.PagingAndSortingRepository;
  * By Anton Krylov (anthony.kryloff@gmail.com)
  * Date: 3/23/19 2:50 PM
  */
-public interface BookRepository extends PagingAndSortingRepository<Book, Long> {
+public interface BookRepository extends PagingAndSortingRepository<Book, Integer> {
 
     @Query("update Book set readAlready = true where id = :bookId and readAlready = false ")
     @Modifying
-    boolean markRead(long bookId);
+    boolean markRead(int bookId);
 
     //todo FTS
     Page<Book> findAllByDescriptionContainsIgnoreCase(String query, Pageable pageable);
